Refactor: split parser and interpreter into logical components
- Split parser/parser.go into parser.go, parser_stmt.go, and parser_expr.go.
- Split interp/state.go into state.go, eval_stmt.go, and eval_expr.go.
- Created common test utilities for parser and interpreter.
- Modernized and split tests for better maintainability.
- Verified all tests pass.
