+++ parser/parser_expr.go
@@ -1,0 +1,503 @@
+package parser
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/fj68/vvlang/ast"
+	"github.com/fj68/vvlang/lexer"
+)
+
+func (p *Parser) parseFunLiteralExpr() (ast.Expr, error) {
+	if p.curToken.Type == lexer.TFun {
+		if err := p.readToken(); err != nil {
+			return nil, err
+		}
+	}
+	if err := p.expect(lexer.TLParen); err != nil {
+		return nil, err
+	}
+
+	args, err := p.parseFunLiteralArgs()
+	if err != nil {
+		return nil, err
+	}
+
+	body, err := p.parseBody()
+	if err != nil {
+		return nil, err
+	}
+
+	if err := p.expect(lexer.TEnd); err != nil {
+		return nil, err
+	}
+
+	return &ast.FunLiteralExpr{
+		Args: args,
+		Body: body,
+	}, nil
+}
+
+func (p *Parser) parseFunLiteralArgs() ([]string, error) {
+	var args []string
+	for {
+		if p.curToken.Type == lexer.TRParen {
+			break
+		}
+		name := p.curToken.Text
+		if err := p.expect(lexer.TIdent); err != nil {
+			return nil, err
+		}
+		args = append(args, name)
+		if p.curToken.Type == lexer.TRParen {
+			break
+		}
+		if err := p.expect(lexer.TComma); err != nil {
+			return nil, err
+		}
+	}
+	if err := p.expect(lexer.TRParen); err != nil {
+		return nil, err
+	}
+	return args, nil
+}
+
+func (p *Parser) parseExpr(precedence Precedence) (expr ast.Expr, err error) {
+	prefix, ok := p.prefixParsers[p.curToken.Type]
+	if !ok {
+		return nil, fmt.Errorf("no prefix parser found for %s", p.curToken.Type)
+	}
+	expr, err = prefix()
+	if err != nil {
+		return nil, err
+	}
+
+	stopTokens := []lexer.TokenType{
+		lexer.TEOF,
+		lexer.TEnd,
+	}
+	for !oneOf(stopTokens, p.curToken.Type) && precedence < precedenceOf(p.curToken.Type) {
+		infix, ok := p.infixParsers[p.curToken.Type]
+		if !ok {
+			break
+		}
+		expr, err = infix(expr)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return expr, nil
+}
+
+func (p *Parser) parsePrefixExpr() (ast.Expr, error) {
+	op := p.curToken.Text
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	right, err := p.parseExpr(PPrefix)
+	if err != nil {
+		return nil, err
+	}
+	return &ast.PrefixExpr{
+		Op:    op,
+		Right: right,
+	}, nil
+}
+
+func (p *Parser) parseInfixExpr(left ast.Expr) (ast.Expr, error) {
+	op := p.curToken.Text
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	right, err := p.parseExpr(PPrefix)
+	if err != nil {
+		return nil, err
+	}
+	return &ast.InfixExpr{
+		Op:    op,
+		Left:  left,
+		Right: right,
+	}, nil
+}
+
+func (p *Parser) parseVarRefExpr() (ast.Expr, error) {
+	name := p.curToken.Text
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.VarRefExpr{Name: name}, nil
+}
+
+func (p *Parser) parseFunCallExpr(fun ast.Expr) (ast.Expr, error) {
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	args, err := p.parseFunCallArgs()
+	if err != nil {
+		return nil, err
+	}
+	return &ast.FunCallExpr{
+		Fun:  fun,
+		Args: args,
+	}, nil
+}
+
+func (p *Parser) parseFunCallArgs() ([]ast.Expr, error) {
+	var args []ast.Expr
+	for {
+		if p.curToken.Type == lexer.TEOF {
+			return nil, fmt.Errorf("unexpected token while reading arguments for function call")
+		}
+		if p.curToken.Type == lexer.TRParen {
+			break
+		}
+		arg, err := p.parseExpr(PLowest)
+		if err != nil {
+			return nil, err
+		}
+		args = append(args, arg)
+		if p.curToken.Type == lexer.TRParen {
+			break
+		}
+		if err := p.expect(lexer.TComma); err != nil {
+			return nil, err
+		}
+	}
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return args, nil
+}
+
+func (p *Parser) parseIndexOrSliceExpr(left ast.Expr) (ast.Expr, error) {
+	// current token is TLBrace ('['
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+
+	// Check for slice notation (with optional start)
+	// or index notation
+	var start ast.Expr
+	var end ast.Expr
+
+	// If we see a colon right away, start is nil and we have a slice like [:end]
+	switch p.curToken.Type {
+	case lexer.TColon:
+		// start is nil
+	case lexer.TRBrace:
+		// Empty brackets [] - error
+		return nil, fmt.Errorf("expected index or slice expression")
+	default:
+		// Parse the first expression
+		expr, err := p.parseExpr(PLowest)
+		if err != nil {
+			return nil, err
+		}
+
+		// Check if it's a slice or index
+		if p.curToken.Type == lexer.TColon {
+			start = expr
+		} else {
+			// It's an index access
+			if p.curToken.Type != lexer.TRBrace {
+				return nil, fmt.Errorf("expected ']' after index, got %s", p.curToken.Type)
+			}
+			if err := p.readToken(); err != nil {
+				return nil, err
+			}
+			return &ast.IndexExpr{
+				Left:  left,
+				Index: expr,
+			}, nil
+		}
+	}
+
+	// Parse the end expression (for slice)
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	if p.curToken.Type != lexer.TRBrace {
+		expr, err := p.parseExpr(PLowest)
+		if err != nil {
+			return nil, err
+		}
+		end = expr
+	}
+
+	if p.curToken.Type != lexer.TRBrace {
+		return nil, fmt.Errorf("expected ']' after slice, got %s", p.curToken.Type)
+	}
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+
+	return &ast.SliceExpr{
+		Left:  left,
+		Start: start,
+		End:   end,
+	}, nil
+}
+
+func (p *Parser) parseNullLiteralExpr() (ast.Expr, error) {
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.NullLiteralExpr{}, nil
+}
+
+func (p *Parser) parseBuiltinCallExpr() (ast.Expr, error) {
+	op := p.curToken.Text
+	pos := p.curToken.Pos
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	if err := p.expect(lexer.TLParen); err != nil {
+		return nil, err
+	}
+	expr, err := p.parseExpr(PLowest)
+	if err != nil {
+		return nil, err
+	}
+	if err := p.expect(lexer.TRParen); err != nil {
+		return nil, err
+	}
+	return &ast.BuiltinCallExpr{
+		Position: ast.Position{Start: &pos, End: &p.curToken.Pos},
+		Op:       op,
+		Value:    expr,
+	}, nil
+}
+
+func (p *Parser) parseInterpolatedStringLiteralExpr() (ast.Expr, error) {
+	pos := p.curToken.Pos
+	text := p.curToken.Text
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+
+	texts := []string{}
+	values := []ast.Expr{}
+
+	runes := []rune(text)
+	last := 0
+	var currentText strings.Builder
+
+	for i := 0; i < len(runes); i++ {
+		if runes[i] == '{' {
+			if i+1 < len(runes) && runes[i+1] == '{' {
+				currentText.WriteString(string(runes[last:i]))
+				currentText.WriteRune('{')
+				i++
+				last = i + 1
+				continue
+			}
+			// Start of interpolation
+			currentText.WriteString(string(runes[last:i]))
+			texts = append(texts, currentText.String())
+			currentText.Reset()
+
+			i++
+			start := i
+			braceCount := 1
+			for i < len(runes) && braceCount > 0 {
+				switch runes[i] {
+				case '{':
+					braceCount++
+				case '}':
+					braceCount--
+				}
+				if braceCount > 0 {
+					i++
+				}
+			}
+			if braceCount > 0 {
+				return nil, fmt.Errorf("missing closing brace in interpolated string")
+			}
+			exprText := string(runes[start:i])
+
+			subParser := New([]rune(exprText))
+			subParser.readToken()
+			subParser.readToken()
+			subExpr, err := subParser.parseExpr(PLowest)
+			if err != nil {
+				return nil, err
+			}
+			values = append(values, subExpr)
+			last = i + 1
+		} else if runes[i] == '}' {
+			if i+1 < len(runes) && runes[i+1] == '}' {
+				currentText.WriteString(string(runes[last:i]))
+				currentText.WriteRune('}')
+				i++
+				last = i + 1
+				continue
+			}
+			// Lone } is just a character if not closing an interpolation
+			// (but our loop only sees this if it's NOT inside one)
+		}
+	}
+	currentText.WriteString(string(runes[last:]))
+	texts = append(texts, currentText.String())
+
+	return &ast.InterpolatedStringLiteralExpr{
+		Position: ast.Position{Start: &pos, End: &p.curToken.Pos},
+		Texts:    texts,
+		Values:   values,
+	}, nil
+}
+
+func (p *Parser) parseDigitLiteralExpr() (ast.Expr, error) {
+	value, err := strconv.ParseFloat(p.curToken.Text, 64)
+	if err != nil {
+		return nil, err
+	}
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.NumberLiteralExpr{
+		Value: value,
+	}, nil
+}
+
+func (p *Parser) parseBoolLiteralExpr() (ast.Expr, error) {
+	value := p.curToken.Type == lexer.TTrue
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.BoolLiteralExpr{
+		Value: value,
+	}, nil
+}
+
+func (p *Parser) parseStringLiteralExpr() (ast.Expr, error) {
+	value := p.curToken.Text
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.StringLiteralExpr{
+		Value: value,
+	}, nil
+}
+
+func (p *Parser) parseListLiteralExpr() (ast.Expr, error) {
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+
+	var elements []ast.Expr
+	for {
+		if p.curToken.Type == lexer.TEOF {
+			return nil, fmt.Errorf("unexpected EOF while reading list")
+		}
+		if p.curToken.Type == lexer.TRBrace {
+			break
+		}
+
+		elem, err := p.parseExpr(PLowest)
+		if err != nil {
+			return nil, err
+		}
+		elements = append(elements, elem)
+
+		if p.curToken.Type == lexer.TRBrace {
+			break
+		}
+
+		if err := p.expect(lexer.TComma); err != nil {
+			return nil, err
+		}
+	}
+
+	if err := p.expect(lexer.TRBrace); err != nil {
+		return nil, err
+	}
+
+	return &ast.ListLiteralExpr{
+		Elements: elements,
+	}, nil
+}
+
+func (p *Parser) parseRecordLiteralExpr() (ast.Expr, error) {
+	// current token is TLBracket ("{")
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	fields := map[string]ast.Expr{}
+	// empty record
+	if p.curToken.Type == lexer.TRBracket {
+		if err := p.readToken(); err != nil {
+			return nil, err
+		}
+		return &ast.RecordLiteralExpr{Fields: fields}, nil
+	}
+	for {
+		if p.curToken.Type != lexer.TIdent {
+			return nil, fmt.Errorf("expected identifier for record field, got %s", p.curToken.Type)
+		}
+		name := p.curToken.Text
+		if err := p.expectNext(lexer.TAssign); err != nil {
+			return nil, err
+		}
+		if err := p.readToken(); err != nil {
+			return nil, err
+		}
+		expr, err := p.parseExpr(PLowest)
+		if err != nil {
+			return nil, err
+		}
+		fields[name] = expr
+		if p.curToken.Type == lexer.TRBracket {
+			break
+		}
+		// allow optional trailing comma before the closing brace
+		if p.curToken.Type == lexer.TComma {
+			// consume comma
+			if err := p.readToken(); err != nil {
+				return nil, err
+			}
+			// if next token is closing brace, break (trailing comma)
+			if p.curToken.Type == lexer.TRBracket {
+				break
+			}
+			// otherwise continue to next field
+			continue
+		}
+		return nil, fmt.Errorf("expected comma or '}', got %s", p.curToken.Type)
+	}
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.RecordLiteralExpr{Fields: fields}, nil
+}
+
+func (p *Parser) parseFieldAccessExpr(record ast.Expr) (ast.Expr, error) {
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	if p.curToken.Type != lexer.TIdent {
+		return nil, fmt.Errorf("expected identifier after '.', got %s", p.curToken.Type)
+	}
+	field := p.curToken.Text
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	return &ast.FieldAccessExpr{
+		Record: record,
+		Field:  field,
+	}, nil
+}
+
+func (p *Parser) parseGroupedExpr() (ast.Expr, error) {
+	if err := p.readToken(); err != nil {
+		return nil, err
+	}
+	expr, err := p.parseExpr(PLowest)
+	if err != nil {
+		return nil, err
+	}
+	if err := p.expect(lexer.TRParen); err != nil {
+		return nil, err
+	}
+	return expr, nil
+}
